requirements.txt
aniso8601==9.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
DateTime==4.9
defusedxml==0.7.1
ecdsa==0.18.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-MySQL==1.5.2
Flask-OpenID==1.3.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
oauthlib==3.2.2
packaging==22.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-core==4.3.0
SQLAlchemy==1.4.45
twilio==7.16.2
urllib3==1.26.14
Werkzeug==2.2.2
zope.interface==5.5.2
